add LinearSolve
