+++ Project.toml
@@ -8,6 +8,7 @@ DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
 LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+LinearSolve = "7ed4a6bd-45f5-4d41-b270-4a48e9bafcae"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 ProgressLogging = "33c8b6b6-d38a-422a-b730-caa89a2f386c"
 SimpleWeightedGraphs = "47aef6b3-ad0c-573a-a1e2-d07658019622"
@@ -19,6 +20,7 @@ ArnoldiMethod = "0.0.4, 0.4"
 DelimitedFiles = "1"
 Graphs = "1"
 LaTeXStrings = "1.1"
+LinearSolve = "2.38.0"
 Plots = "1.4"
 ProgressLogging = "0.1"
 SimpleWeightedGraphs = "1.1"
