+++ src/problem.jl
@@ -4,21 +4,20 @@ abstract type RSPOperation <: AbstractOperation end
 abstract type AbstractProblem end
 
 """
-    compute(o, grid::Union{Grid,GridRSP})
+    compute(problem, grid::Union{Grid,GridRSP})
 
-Compute operation `o` on a grid.
+Compute problem `o` on a grid.
 """
 function compute end
 
 """
-    assess(o::Operation, g::Grid)
+    assess(p::AbstractProblem, g)
 
-Assess the memory and compute requirements of operation
-`o` on grid `g`. This can be used to indicate memory
+Assess the memory and compute requirements of problem
+`p` on grid `g`. This can be used to indicate memory
 and time reequiremtents on a cluster
 """
 function assess end
-
 
 """
     SolverMode
@@ -30,6 +29,9 @@ abstract type SolverMode end
    MaterializedSolve()
 
 Compute all operations on a fully materialised Z matrix.
+
+This is inneficient for CPUS, but may be best for GPUs
+using CuSSP.jl
 """
 struct MatrixSolve <: SolverMode end
 """
@@ -37,9 +39,10 @@ struct MatrixSolve <: SolverMode end
 
 Compute all operations column by column,
 after precomputing lu decompositions.
+
+TODO we can put solver details 
 """
 struct VectorSolve <: SolverMode end
-
 
 """
     Problem(ops...; mode, θ)
@@ -48,7 +51,7 @@ Combine multiple compute operations into a single object,
 to be run in the same job.
 """
 @kwdef struct Problem{M<:SolverMode,O,T} <: AbstractProblem
-    mode::M=MaterializedSolve()
+    mode::M=VectorSolve()
     ops::O
     θ::T=nothing
 end
@@ -56,126 +59,55 @@ Problem(args...; kw...) = Problem(args; kw...)
 Problem(p::AbstractProblem; θ=nothing, mode=p.mode) = Problem(o.ops; mode, θ)
 
 compute(p::Problem, g::Grid) = compute(p.mode, p, g::Grid)
+
 # Use an iterative solver so the grid is not materialised
 function compute(m::VectorSolve, p::Problem, g::Grid)
+    # Compute Z column by column
+    _, targetnodes = _targetidx_and_nodes(g)
     P = _Pref(g.affinities)
     W = _W(P, θ, g.costmatrix)
     # Sparse lhs
-    a = I - W
+    A = I - W
+    b = sparse(targetnodes,
+        1:length(targetnodes),
+        1.0,
+        size(g.costmatrix, 1),
+        length(targetnodes),
+    )
     # Dense rhs column
     b = zeros(eltype(a), size(g.affinities, 2))
     # Define and initialise the linear problem
-    linprob = LinearProblem(a, b)
+    linprob = LinearProblem(A, b)
     linsolve = init(linprob)
-    map(columns) do 
-        fill!(b, zero(eltype(b)))
-        # fill b with column val
+    map(1:size(a, 2)) do i 
+        b .= view(A, i)
+        # Update solver with new b values
         linsolve = LinearSolve.set_b(linsolve, b)
         sol = solve(linsolve)
         res = sol2.u
-        compute(o, res)
+        # compute(p, res)
     end
 end
 # Materialise the whole rhs matrix
 function compute(m::MatrixSolve, o::Problem, g::Grid) 
-    # We could just move all the GridRSP code here?
-    grsp = GridRSP(g; allocs=o.allocs)
+    # Legacy code... but maybe materialising is faster for CUDSS?
+    Pref = _Pref(g.affinities)
+    W    = _W(Pref, θ, g.costmatrix)
+    targetidx, targetnodes = _targetidx_and_nodes(g)
+    Z    = (I - W) \ Matrix(sparse(targetnodes,
+                                 1:length(targetnodes),
+                                 1.0,
+                                 size(g.costmatrix, 1),
+                                 length(targetnodes)))
+    # Check that values in Z are not too small:
+    if minimum(Z) * minimum(nonzeros(g.costmatrix .* W)) == 0
+        @warn "Warning: Z-matrix contains too small values, which can lead to inaccurate results! Check that the graph is connected or try decreasing θ."
+    end
+
     map(o.ops) do op
         compute(m, op, grsp)
     end
 end
-
-
-"""
-    WindowedProblem(op; size, centers, θ)
-
-Combine multiple compute operations into a single object, 
-to be run over the same windowed grids.
-
-"""
-@kwdef struct WindowedProblem{O,SS,WS,WC} <: AbstractProblem
-    op::O
-    sourcesize::SS
-    size::WS
-    centers::WC
-end
-function WindowedProblem(op::O, radius::Tuple, windowcenters::AbstractMatrix; 
-    θ::T=nothing
-) where O<:Tuple 
-    WindowedProblem{O}(Problem(op; θ), windowsize, windowcenters)
-end
-
-function compute(p::WindowedProblem, rasterstack)
-    map(_window_grids(p, rasterstack)) do g
-        compute(p, g)
-    end
-end
-
-
-"""
-    TiledProblem(op; size, centers, θ)
-
-Combine multiple compute operations into a single object, 
-to be run over tiles of windowed grids.
-
-"""
-@kwdef struct TiledProblem{O,SS,WS,WC}
-    op::O
-    ranges::R
-    layout::M
-end
-function TiledProblem(w::WindowedProblem;
-    target::Raster,
-    radius::Real,
-    overlap::Real,
-) where O<:Tuple 
-    res = resolution(target)
-    # Convert distances to pixels
-    r = radius / res
-    o = overlap / res
-    s = r - o # Step between each tile corner
-    # Get the corners of each tile
-    ci = CartesianIndices(target)[begin:s:end, begin:s:end]
-    # Create an array of ranges for retreiving each tile
-    tile_ranges = map(ci) do tile_corner
-        map(tile_corner, size(target)) do i, sz
-            i:min(sz, i + r)
-        end
-    end
-    # Create a mask to skip tiles that have no target cells
-    mask = map(ci) do tile_starts
-        # Retrive a tile
-        tile = view(target, tile_ranges...)
-        # If there are non-NaN cells above zero, keep the tile
-        any(x -> !isnan(x) && x > zero(x), tile)
-    end
-
-    return TiledProblem(w, ranges, mask)
-end
-
-function compute(p::TiledProblem, rast)
-    map(p.ranges, p.mask) do rs, m
-        m || return nothing
-        tile = rast[rs...]
-        g = Grid(tile)
-        outputs = compute(p, w)
-        write(outputs, p.storage)
-        nothing
-    end
-end
-
-# function assess(op::WindowedProblem, g::Grid) 
-#     window_assessments = map(_windows(op, g)) do w
-#         ca = assess(op.op, w)
-#     end
-#     maximums = reduce(window_assessments) do acc, a
-#         (; totalmem=max(acc.totalmem, a.totalmem),
-#            zmax=max(acc.zmax, a.zmax),
-#            lumax=max(acc.lumax, a.lumax),
-#         )
-#     end
-#     ComputeAssesment(; op=op.op, maximums..., sums...)
-# end
 
 # @kwdef struct ComputeAssesment{P,M,T}
 #     problem::P
